Unify auto chart generation across result profiles
